Fixed SIFD-89: Crash when no activity found

+++ app/models/instance.rb
@@ -54,6 +54,8 @@ class Instance < ActiveFedora::Base
   end
 
   def content_files=(files)
+    # ensure instance is valid before saving files
+    return unless self.valid?
     #remove old file
     content_files.delete_all
     files.each do |f|
